Add tool to simulate a snake draft many times and report results.
The main idea is to answer the question of should I draft
WR-WR or RB-RB? The program draft.py will randomly assign
positional player draft orders to teams (e.g. WR-RB-TE-WR-QB-RB)
and execute those orders in a snake style draft. When a team
drafts a particular position, the next best player at that
position is selected.

Results are aggregated by the team's draft position by
cumulative team projected points. Naturally, the higher the
point total, the better the draft went for that team.