Add basic auction draft simulator.
In the auction draft, teams nominate players based on
random order of the positions in their roster. The resulting
auction is effectively a silent auction with teams bidding their
max budget for the player at the position regardless of what the
other teams bid. The team with the winning bid drafts the auctioned
player.